chore(ci): towards ci3 (#10751)
CI3 is a conceptual goal for uniting the CI flow and the dev flow as
much as possible, adding more depth to the bootstrap and build scripts
to be able to handle our needs.

This PR introduces all the work on CI3 so far, but still has an earthly
caller shell to make sure we can minimize the number of variables that
have changed at once.

There is a lot of changes in this PR.
See https://github.com/AztecProtocol/aztec-packages/pull/10711/files for
a subset of the changes without yarn.lock etc noise.

The big picture:
- The CI build has been made much less stateful. ci.yml now uses the ci3
bootstrap pattern, without fully moving off the earthly targets just
yet.
- The S3 cache mechanism is now the main cache mechanism. Note there is
no persistent disk now supporting the build.
There is a global cache on S3, readable without auth, that caches them
for 10 days. We no longer think of the build in terms of docker/buildkit
layers but instead as chunks that have different rebuild patterns that
match files in the monorepo.
- Moving to yarn 4.5.2. 

Niceties:
- faster builds due to script improvements and distributed cache
uploading by default
- work is more properly isolated in chunks from the above effort
- spot recovery is implemented, retrying with on-demand
- we no longer use github runners, side-stepping lots of edge-cases, and
instead rely on our builder realizing there is no work to do / hitting a
timeout via shutdown -P
- Docker images are no longer copied from the builder, meaning a large
class of flake is gone.

Non-niceties:
- The earthly setup is much less granular. There is two stages that have
their own one-layer builds. The earthly cache is fairly redundant, using
the S3 cache for most meaningful caching. (earthly will not be used in
ci.yml in the future)
- Some CI files are now duplicated, we will do a follow-on pass to get
rid of earthly helpers, build-system, etc
- CI currently also downloads the CI image fresh each time, will change
- we are currently pushing images to dockerhub with no expiration,
should move to ECR
- noir-projects currently retries once in the Earthfile as a last minute
issue was hit, will be fixed in a follow-up
- Issues: https://github.com/AztecProtocol/aztec-packages/issues/10775

---- 
WORKFLOW AFTER THIS PR:
- Run ./bootstrap.sh in root to bootstrap with cache, ./bootstrap.sh
full otherwise
- Run earthly +ci in root to 
- Put ci3 in your cache and note the ways to interact with ci in that
folder
- Note the new commands in ./bootstrap.sh like test-kind-network
- In yarn-project to run a single e2e test now use `test:e2e`. `test`
just runs the unit tests as per other projects.

---------

+++ barretenberg/acir_tests/bootstrap.sh
@@ -5,11 +5,13 @@ cmd=${1:-}
 export CRS_PATH=$HOME/.bb-crs
 
 function build {
-  # Clone the execution success tests from noir repo.
-  rm -rf acir_tests
-  cp -R ../../noir/noir-repo/test_programs/execution_success acir_tests
-  # Running these requires extra gluecode so they're skipped.
-  rm -rf acir_tests/{diamond_deps_0,workspace,workspace_default_member}
+  if [ ! -d acir_tests ]; then
+    cp -R ../../noir/noir-repo/test_programs/execution_success acir_tests
+    # Running these requires extra gluecode so they're skipped.
+    rm -rf acir_tests/{diamond_deps_0,workspace,workspace_default_member}
+    # TODO(https://github.com/AztecProtocol/barretenberg/issues/1108): problem regardless the proof system used
+    rm -rf acir_tests/regression_5045
+  fi
 
   # COMPILE=2 only compiles the test.
   github_group "acir_tests compiling"
@@ -38,9 +40,14 @@ function build {
   github_endgroup
 }
 
+function hash {
+  cache_content_hash ../../noir/.rebuild_patterns ../../noir/.rebuild_patterns_tests ../../barretenberg/cpp/.rebuild_patterns ../../barretenberg/ts/.rebuild_patterns
+}
 function test {
-  local hash=$(cache_content_hash ../../noir/.rebuild_patterns_native ../../noir/.rebuild_patterns_tests ../../barretenberg/cpp/.rebuild_patterns ../../barretenberg/ts/.rebuild_patterns)
+  github_group "acir_tests testing"
+  local hash=$(hash)
   if ! test_should_run barretenberg-acir-tests-$hash; then
+    github_endgroup
     return
   fi
 
@@ -139,6 +146,9 @@ case "$cmd" in
   "ci")
     denoise build
     denoise test
+    ;;
+  "hash")
+    hash
     ;;
   "test")
     denoise test
